Revert "Use upstream tile, fuse, and distribution for codegen driver approaches. (#8614)"
This reverts commit 16e6dabd83fe56efdae869d41b5dfe4cc2820f2b.
