+++ iree/compiler/Codegen/Common/BUILD
@@ -51,7 +51,6 @@ cc_library(
         "OptimizeVectorTransferPass.cpp",
         "PolynomialApproximationPass.cpp",
         "RemoveTrivialLoops.cpp",
-        "RewriteLinalgDestructiveUpdatesPass.cpp",
         "SetNumWorkgroupsPass.cpp",
         "TileAndDistributeToWorkgroupsPass.cpp",
         "TypePropagationPass.cpp",
