+++ iree/compiler/Codegen/Common/CMakeLists.txt
@@ -43,7 +43,6 @@ iree_cc_library(
     "OptimizeVectorTransferPass.cpp"
     "PolynomialApproximationPass.cpp"
     "RemoveTrivialLoops.cpp"
-    "RewriteLinalgDestructiveUpdatesPass.cpp"
     "SetNumWorkgroupsPass.cpp"
     "TileAndDistributeToWorkgroupsPass.cpp"
     "TypePropagationPass.cpp"
