+++ iree/compiler/Codegen/Common/test/BUILD
@@ -34,7 +34,6 @@ iree_lit_test_suite(
             "linalg_bufferize.mlir",
             "remove_dead_allocs.mlir",
             "remove_trivial_loops.mlir",
-            "rewrite_linalg_destructive_updates.mlir",
             "tile_and_distribute_to_workgroups.mlir",
             "transpose_canonicalization.mlir",
             "type_propagation.mlir",
