+++ iree/compiler/Codegen/Sandbox/Passes.h
@@ -7,7 +7,6 @@
 #ifndef IREE_CODEGEN_SANDBOX_PASSES_H_
 #define IREE_CODEGEN_SANDBOX_PASSES_H_
 
-#include "mlir/Dialect/Linalg/Utils/Utils.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
@@ -25,7 +24,6 @@ struct LinalgFusePassOptions {
   bool vectorize = false;
   bool vectorizePadding = false;
   int64_t tilingLevel = -1;
-  bool doIREEDistribution = false;
 };
 
 /// Creates a pass to drive tile + fuse transformations of `LinalgOp`s.
