+++ iree/compiler/Codegen/Sandbox/Passes.td
@@ -43,9 +43,7 @@ def LinalgFuse : Pass<"linalg-fuse", "FuncOp"> {
 
     // IREE specific options
     Option<"tilingLevel", "tiling-level", "int64_t", /*default=*/"-1",
-      "Use default tiling level used to retrieve the configuration from lowering_config">,
-    Option<"doIREEDistribution", "do-iree-distribution", "bool", /*default=*/"false",
-      "Apply IREE distribution during tiling">,
+      "Use default tiling level used to retrieve the configuration from lowering_config">
   ];
   let dependentDialects = [
     "::mlir::arith::ArithmeticDialect", "::mlir::AffineDialect",
